+++ R/pdf_add_crosshairs.R
@@ -58,17 +58,48 @@ pdf_add_crosshairs <- function(input, output = NULL, ...,
         r <- pdf_render_raster(input, page = i, dpi = dpi)
         grid.raster(r, interpolate = FALSE, vp = vp)
 
-        pixmap <- pdf_render_bm_pixmap(input, page = i, dpi = dpi)
         if (i %in% pages) {
-            for (j in seq_len(nrow(layout))) {
-                piecepackr::grid.crosshair(
-                    x = layout$x[j], y = layout$y[j],
-                    width = layout$width[j], height = layout$height[j],
-                    default.units = "in", ...
-                )
-            }
+            grid_add_crosshairs(..., layout = layout)
         }
     }
     invisible(dev.off())
     invisible(output)
+}
+
+#' Draw (rounded) rectangles around components
+#'
+#' `grid_add_rects()` draws (rounded) rectangles around components of a print-and-play layout.
+#'
+#' * This function draws in **inches** so make sure your graphics device is "big" enough.
+#'
+#' @inheritParams pdf_pad_paper
+#' @inheritParams pdf_pages
+#' @inheritParams bm_crop_layout
+#' @param ... Passed to [piecepackr::grid.crosshair()].
+#' @return `NULL` invisibly.
+#'         As a side effect draws crosshairs to the active graphics device.
+#' @examples
+#' if (requireNamespace("piecepackr", quietly = TRUE) &&
+#'     utils::packageVersion("piecepackr") >= "1.14.0-5") {
+#'   grid::grid.newpage()
+#'   vp <- grid::viewport(width=8.5, height=11, default.units="in",
+#'                        x=0.5, y=0.5, just=c("left", "bottom"))
+#'   grid::pushViewport(vp)
+#'   grid_add_crosshairs(layout = "poker_3x3")
+#'   grid::popViewport()
+#' }
+#' @export
+grid_add_crosshairs <- function(..., layout = "poker_3x3") {
+    stopifnot(requireNamespace("piecepackr", quietly = TRUE))
+    stopifnot(packageVersion("piecepackr") >= "1.14.0-5")
+    if (is.character(layout))
+        layout <- layout_preset(layout)
+    for (j in seq_len(nrow(layout))) {
+        piecepackr::grid.crosshair(
+            x = layout$x[j], y = layout$y[j],
+            width = layout$width[j], height = layout$height[j],
+            default.units = "in", ...
+        )
+    }
+    invisible(NULL)
 }