+++ tests/testthat/test-pdf_rm_crosshairs.R
@@ -1,6 +1,6 @@
 test_that("`pdf_rm_crosshairs()`", {
     skip_if_not_installed("bittermelon")
-    skip_if_not_installed("piecepackr", "1.14.0-5")
+    skip_if_not_installed("piecepackr", "1.14.0-6")
     on.exit(rm_temp_pdfs(), add = TRUE)
 
     f1 <- pdf_create_blank(length = 2L, width = 11, height = 8.5)